Seo-integration: refactor setup
- move replacement variables to own file
- move seo title width to own file
- move registering to index: this makes the app a component without import side-effects
